+++ raycasting/raycasting_utils.c
@@ -6,7 +6,7 @@
 /*   By: zbakour <zbakour@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/12 21:24:32 by zbakour           #+#    #+#             */
-/*   Updated: 2025/06/12 21:25:32 by zbakour          ###   ########.fr       */
+/*   Updated: 2025/06/21 17:27:47 by zbakour          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,21 +14,21 @@
 
 void define_player_angle(t_game *game)
 {
-	if (game->player_char == 'N')
-		game->player_angle = M_PI / 2; 
-	else if (game->player_char == 'S')
-		game->player_angle = 3 * M_PI / 2; 
-	else if (game->player_char == 'E')
-		game->player_angle = 0.0; 
-	else if (game->player_char == 'W')
-		game->player_angle = M_PI;
+    if (game->player_char == 'N')
+        game->player_angle = M_PI / 2;
+    else if (game->player_char == 'S')
+        game->player_angle = 3 * M_PI / 2;
+    else if (game->player_char == 'E')
+        game->player_angle = 0;
+    else if (game->player_char == 'W')
+        game->player_angle = M_PI;
 }
 
-double normalize_angle(float angle)
+double normalize_angle(double angle)
 {
-    while (angle < 0)
-        angle += 2 * M_PI;
-    while (angle >= 2 * M_PI)
-        angle -= 2 * M_PI;
-    return angle;
+	while (angle < 0)
+		angle += 2 * M_PI;
+	while (angle >= 2 * M_PI)
+		angle -= 2 * M_PI;
+	return angle;
 }
